requirements.txt
altair==4.2.0
numpy==1.22.4
pandas==1.4.2
pydeck==0.7.1
seaborn
matplotlib
streamlit-aggrid==0.2.3.post2
plotly==5.9.0
streamlit_echarts==0.4.0
streamlit-authenticator==0.1.1
pyarrow==8.0.0
python_dateutil==2.8.2
requests==2.28.1
scikit_learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
torch==1.12.0
./streamlit-1.9.0-py2.py3-none-any.whl
